requirements.txt
flet
pyserial
requests